+++ src/scancode/cli_test_utils.py
@@ -17,6 +17,7 @@ from commoncode.system import on_windows
 from packageurl import PackageURL
 
 from scancode_config import scancode_root_dir
+from scancode_config import REGEN_TEST_FIXTURES
 
 
 def run_scan_plain(
@@ -405,3 +406,32 @@ def streamline_jsonlines_scan(scan_result, remove_file_date=False):
 
         for scanned_file in result_line.get('files', []):
             streamline_scanned_file(scanned_file, remove_file_date)
+
+
+def check_json(expected, results, regen=REGEN_TEST_FIXTURES):
+    """
+    Assert if the results JSON file is the same as the expected JSON file.
+    """
+    if regen:
+        mode = 'w'
+        with open(expected, mode) as ex:
+            json.dump(results, ex, indent=2, separators=(',', ': '))
+    with open(expected) as ex:
+        expected = json.load(ex)
+    assert results == expected
+
+
+def load_both_and_check_json(expected, results, regen=REGEN_TEST_FIXTURES):
+    """
+    Assert if the results JSON file is the same as the expected JSON file.
+    """
+    with open(results) as res:
+        results = json.load(res)
+
+    if regen:
+        mode = 'w'
+        with open(expected, mode) as ex:
+            json.dump(results, ex, indent=2, separators=(',', ': '))
+    with open(expected) as ex:
+        expected = json.load(ex)
+    assert results == expected